feat  : pakjungmin namespace 적용
