--- Assets/PAKJUNGMIN/01.Scripts/BaseTile.cs
@@ -1,16 +1,0 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEditor.Experimental.GraphView;
-using UnityEngine;
-
-public class BaseTile : MonoBehaviour
-{
-    public Node tileNode;
-
-}
-
-public struct Pos
-{
-
-}
