--- Assets/PAKJUNGMIN/01.Scripts/Bomb.cs
@@ -1,17 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Bomb : MonoBehaviour
-{
-    [SerializeField] Transform bombPos;
-
-    private void OnEnable()
-    {
-        
-    }
-    void Explode()
-    {
-
-    }
-}
