+++ Assets/PAKJUNGMIN/01.Scripts/Item.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 81f86609ab842944ab433549460a6ccc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
