+++ Assets/PAKJUNGMIN/01.Scripts/Item/Bomb.cs
@@ -1,0 +1,21 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+
+namespace pakjungmin
+{
+    public class Bomb : MonoBehaviour
+    {
+        [SerializeField] Transform bombPos;
+
+        private void OnEnable()
+        {
+
+        }
+        void Explode()
+        {
+
+        }
+    }
+}
