+++ Assets/PAKJUNGMIN/01.Scripts/Manager/TileManager.cs
@@ -1,0 +1,24 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace pakjungmin {
+
+    public class TileManager : MonoBehaviour
+    {
+        //Dictionary<BaseTile,Vector2> 
+
+        private void Awake()
+        {
+            //for (int x = 0; x < tileArray.GetLength(0); x++)
+            //{
+            //    for (int y = 0; y < tileArray.GetLength(1); y++)
+            //    {
+            //        //tileArray[x, y].name = $"{x},{y}";
+            //    }
+            //}
+        }
+    }
+
+}
+
