+++ Assets/PAKJUNGMIN/01.Scripts/Map/Tile.cs
@@ -1,0 +1,21 @@
+using UnityEditor.Experimental.GraphView;
+using UnityEngine;
+
+
+
+namespace pakjungmin
+{
+    public class Tile : MonoBehaviour
+    {
+        public Node tileNode;
+
+    }
+
+    public struct Pos
+    {
+
+    }
+
+
+}
+
