+++ Assets/PAKJUNGMIN/01.Scripts/Player/PlayerBehavior.cs
@@ -2,22 +2,27 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class PlayerBehavior : MonoBehaviour
+namespace pakjungmin 
 {
-    PlayerMediator playerMediator;
-   
-    public void Move(Vector3 moveDir) 
-    {
-       // CharacterStats stat = playerMediator.playerStats;
-       // gameObject.transform.Translate(moveDir*stat.Speed,Space.World);
-    }
 
-    public void Plant()
+    public class PlayerBehavior : MonoBehaviour
     {
+        PlayerMediator playerMediator;
 
-    }
-    public void Use()
-    {
+        public void Move(Vector3 moveDir)
+        {
+            // CharacterStats stat = playerMediator.playerStats;
+            // gameObject.transform.Translate(moveDir*stat.Speed,Space.World);
+        }
 
+        public void Plant()
+        {
+
+        }
+        public void Use()
+        {
+
+        }
     }
 }
+
