+++ Assets/PAKJUNGMIN/01.Scripts/Player/PlayerStats.cs
@@ -3,35 +3,39 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class PlayerStats : MonoBehaviour
+
+namespace pakjungmin
 {
-    [SerializeField] PlayerMediator playerMediator;
-
-    OwnStats ownStats;
-
-    public float Speed { get { return ownStats.speedValue; } set { ownStats.speedValue = value; } }
-    public float Bomb { get { return ownStats.bombValue; } set { ownStats.bombValue = value; } }
-    public float Power { get { return ownStats.powerValue; } set { ownStats.powerValue = value; } }
-
-    void Awake()
+    public class PlayerStats : MonoBehaviour
     {
-        InitSet();
+        [SerializeField] PlayerMediator playerMediator;
+
+        OwnStats ownStats;
+
+        public float Speed { get { return ownStats.speedValue; } set { ownStats.speedValue = value; } }
+        public float Bomb { get { return ownStats.bombValue; } set { ownStats.bombValue = value; } }
+        public float Power { get { return ownStats.powerValue; } set { ownStats.powerValue = value; } }
+
+        void Awake()
+        {
+            InitSet();
+        }
+
+        void InitSet()
+        {
+
+            ownStats.powerValue = playerMediator.characterStats.Power;
+            ownStats.speedValue = playerMediator.characterStats.Speed;
+            ownStats.bombValue = playerMediator.characterStats.Bomb;
+        }
     }
 
-    void InitSet()
+
+    [Serializable]
+    public struct OwnStats
     {
-        
-        ownStats.powerValue = playerMediator.characterStats.Power;
-        ownStats.speedValue = playerMediator.characterStats.Speed;
-        ownStats.bombValue = playerMediator.characterStats.Bomb;
+        public float speedValue;
+        public float bombValue;
+        public float powerValue;
     }
-}
-
-
-[Serializable]
-public struct OwnStats
-{
-    public float speedValue;
-    public float bombValue;
-    public float powerValue;
 }
