--- Assets/PAKJUNGMIN/01.Scripts/TileManager.cs
@@ -1,20 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class TileManager : MonoBehaviour
-{
-    //Dictionary<BaseTile,Vector2> 
-
-    private void Awake()
-    {
-        //for (int x = 0; x < tileArray.GetLength(0); x++)
-        //{
-        //    for (int y = 0; y < tileArray.GetLength(1); y++)
-        //    {
-        //        //tileArray[x, y].name = $"{x},{y}";
-        //    }
-        //}
-    }
-}
-
